# Blend images optimizations