This addresses the performance problems [mentioned here](https://github.com/joeyballentine/chaiNNer/issues/984#issuecomment-1250332852).

This PR includes 2 optimizations:
1. Clipping. Most blend mode always produce normalized images (values between 0 and 1). We don't need to call `np.clip` for those blend modes, which is a quite expensive function, as it turns out. In my tests, this saves about 140ms.
2. RGB base. If the base image is an RGB image, then the alpha blending becomes A LOT simpler. In my tests, this saves about 240ms.

In my tests, these 2 optimization together reduced the time it took to blend an RGB base image with an RGBA overlay of the same size from 500ms to 120ms. That's 4x faster.

I also changed the guarantees of the `blend_images` function. It now guarantees to return a normalized image. This not only enables the clipping optimization, it also makes for a cleaner API.